Refactor stripe logic into own ios component
